backend/requirements.txt
aniso8601==8.0.0
attrs==19.3.0
Click==7.0
Flask==1.1.1
flask-restx==0.1.1
importlib-metadata==1.5.0
itsdangerous==1.1.0
Jinja2==2.11.1
jsonschema==3.2.0
MarkupSafe==1.1.1
pkg-resources==0.0.0
psycopg2-binary==2.8.4
pyrsistent==0.15.7
pytz==2019.3
six==1.14.0
Werkzeug==0.16.1
zipp==3.0.0
